Use what is apparently a working match
Prior to this commit, for which I apologise took so long, this script
actually didn't work. After my live stream on YouTube, people reported
that I was still being shown as `Live`, which sucks.

Thanks to 'sanet' on Discord for sharing a script he/she wrote, which
was itself inspired by this one. Using sanet's pattern, maybe we'll
have some luck! YouTube is a pain...

If this works out, I think the next step will be to allow you to select
the channel based on a flag, and with that, have another flag which
shows all, perhaps based on a configuration file in '~/.config'. Ideas,
ideas, ideas...